CUT 'tests/Traits' directory & moved methods in `TestCase`
